requirements.txt
flask==3.0.0
flask-cors==4.0.0
crewai==0.11.0
yt-dlp==2024.3.10
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.116.0
langchain==0.1.0
langchain-openai==0.0.5